fixed imask.c

+++ libs/run_target/odometry_calculate.h
@@ -32,7 +32,7 @@ extern long odometry_current_mm(const odometry_t *odometry, int axis_index);
 
 
 extern long odometry_direction_difference(const odometry_t *odometry,
-                                          long target_direction);
+                                          unsigned short target_direction);
 
 #endif /* !QRK_ODOMETRY_CALCULATE_H */
